Free digital pins 2 and 3 to allow interrupts from buttons
For the state machine to react on button presses, we need to be able to link
the buttons to the pins that generate interrupts. On most boards, and 28-based
arduinos such as Uno or Nano, in particular, pins 2 and 3 are the ones that can
generate interrupts, so we start the digital pin usage of the SensorsAndPump-s
from pin 4, instead of 2.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>