Add an initial menu logic, used in interactive menus
All menus behave identically, so it makes sense to unify on that idea.

One idea is to switch from the current state to wss_menusel on menu
activation (Pressing OK); when in wss_menusel the state machine should
not do transitions directly, but employ the interim active Menu class,
and its state.

This will require the addition of an indirection level in "what's the
nest state?" decision, since, depending on the active menu, the effects
would be different. For now, add some initial pieces necessary for menus
to work.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>