Reorder wss_state members and state machine arrays
GCC seems to implement only trivial designated initalizers, that is, the
designators can appear, but they must be in exact numerical order
without any skips in the sequence, and only the last part of the array
can be skipped to be initialized by default.

This isn't too helpful for writing only the relevant non-implicit
values, or to ignore the original enum order used for indexing, but
at least is going to fail to compile if we have a misalignment between
the arrays, ensuring state machine data consistency.

To reduce the ammount of written defaults we also reorder the enum
states in such a way that the state 0 is wss_sleep, which is the default
initalizer, and also the most likely states to have sleep as a next
state to be placed towards the end of the enum, to allow us to omit the
initalizers for a wss_sleep as the next state as much as possible.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>