Aggressively reduce RAM usage due to DEBUG usage
Use as much as possible DEBUG_P + DEBUG to minimize the strings held in
RAM.

We want to do this since it seems the curren RAM breaking point for
stack/heap/data memory corruption is around 750-800 bytes of the global
data on Arduino Nano.

This is already a little concerning since we still need to implement
GeneROMst, system status, logging and use more than one calibration
point for each module.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>