Show CtrlOne on LCD, reprint status for current in menu_one_x
Implement showing the control plant screen in the control_current menu
states and going back to status, if the control screen is closed.

Even if inefficient, we need to reprint the status for the current
module whenever we enter wss_menu_one_x becasue when returning from ctrl
menu, the screen will contain the control menu screen, instead of the
status screen of the plant.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>