Fix buffer overflow issue introduced in e90fa05e88d7cedc0351bacf97a006e4340c21e1
When dropping the usage of stack allocated buffers in favor of using
_lcd_line* internal scratch buffers, I changed the declaration with
initialization to 0 of the local buffers into a single character write
past the _lcd_line* buffers.

It's ironic my intention was to 0 initialize the entire string buffers to
make sure I had a null terminator just in case some memory safety issue
could truncate the output prematurely.

Rust has me spoiled forever. Damn you C and C++!

To properly fix this we memset to 0 the _lcd_line* buffers, which is
also not ideal since one could easly pass mismatched paramaters to
memset (&foo and sizeof(bar)). Since we don't do funny business with
these buffers, and all buffers are arrays, we can get some safety
with a macro making sure the address and sizeof are matched.

It is also very irnonic that Dan Saks in his “extern c: Talking to C
Programmers about C++” talk starts the C-to-C++ migration with memcpy,
another string.h API that exposes the same problem as our memset. Still,
the original bug was made possible by the way C-style strings work and
the total lack of memory safety of C/C++ in this case.

So, yeah, I'm inreconcilably spoiled by Rust and officialy feeling
friction with C and C++. Unfortunately AVR-Rust is not yet that mature
to work with Arduino libs, at least to my knowledge.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>