Move states with automatic exits at the beginning, test them
States which perform an operation and should leave the state on finish
actually would need a minimal timeout period, but the smallest possible
timeout is also good.

Since automatic-leave states need this timeout and most states use the
default-to-sleep timeout, we move these particular states at the
beginning to have a minimum definition.

Add and test toggle usage state to allow switching on/off the modules.

We also allow theoretical key presses in:
-  manual watering (OK repeats, Next returns) and on finish go mack to
   the menu_one (X) state
- automatic watering, which would go to sleep

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>