WS: Save and restore the enabled state, too (#8)
We don't really care much that my default the previous stored info marks
all modules as disabled. We just need to change that on the system, then
go through "state 25" and the correct enabling info will be set. No need
to bump the layout version and lose the existing calibration info.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>