Sleep aggresivley, keep button responsiveness, and adjust millis
Aggressive sleep times introduces issues wrt button debouncing and time
keeping.

Debouncing
==========
In sleep mode, the isPressed function is called on every loop. If the
idle state is kept at a fixed period, we either come out of sleep way
too often, or button presses seem very sluggish as the minimal debounce
recheck will be the sleep period.

To keep things responsive and also go to low power as much as possible,
we reset to the minimum the low power sleep period on every button ISR
call, while increasing it in wss_sleep until the longest sleep period
(except forever).

We also added a "heart beat" pulse before each sleep cycle, as a nice
side effect of wanting to know the mechanism works well.

Time keeping
============
For sleep, if we don't turn TIMER0 off we'll always sleep for just 1ms,
but if we do, we ruin the the mills() value as TIMER0 is millis() base.

The solution
============
Even if imprecise, just assuming all sleep cycles have been completed
has the advantage no alteration of the LowPower library is needed, plus
is a lot more simpler. Is not a big deal if we water a little more
earlier, and it seems it doesn't break button debouncing.

We assume all sleep cycles were completed, and based on that (clearly
wrong) assumption, adjust our rough estimate of the passed time, so we
do have a roughtly correct autowatering cycle while also keeping the
system in idle for longer periods.

This is indeed inaccurate, but that is fine. The immediate alternative
is to modify the LowPower library to have our own modified ISR handler,
and we would under-estimate the sleep in that case, as incomplete sleep
cycles would not be counted. Also, the WDT timer is inaccurate, so that
also introduces some error.

For our case, we don't care, we're fine to waking up at every 2, 3 or
even 6 hours to autowater, so a few seconds of error is nothing and
there are no hard dead lines for watering.

So this way we do aggressive sleeping (only in wss_sleep state), we keep
responsiveness to button presses even in sleep mode, and we do have a
good enough estimation of the time spent in sleep.

Note: We want to use stock Adrduino boards with least amount of
modifications, but with custom boards we could have used a slower
crystal, or changed fuses to use another time base than the 16MHz on the
Nano.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>