ino,ButtonWS: Add sleep support, but keep responsiveness to button presses
In sleep mode, the isPressed function is called on every loop. If the
idle state is kept at a fixed period, we either come out of sleep way
too often, or button presses seem very sluggish as the minimal debounce
recheck will be the sleep period.

To keep things responsive and also go to low power as much as possible,
we reset to the minimum the low power sleep period on every button ISR
call, while increasing it in wss_sleep until the longest sleep period
(except forever).

We also added a "heart beat" pulse before each sleep cycle, as a nice
side effect of wanting to know the mechanism works well.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>