WIP: Assume all sleep cycles were completed
We assume all sleep cycles were completed, and based on that (clearly
wrong) assumption, adjust our rough estimate of the passed time, so we
do have a roughtly correct autowatering cycle while also keeping the
system in idle for longer periods.

This is indeed inaccurate, but that is fine. The immediate alternative
is to modify the LowPower library to have our own modified ISR handler,
and we would under-estimate the sleep in that case, as incomplete sleep
cycles would not be counted. Also, the WDT timer is inaccurate, so that
also introduces some error.

For our case, we don't care, we're fine to waking up at every 2, 3 or
even 6 hours to autowater, so a few seconds of error is nothing and
there are no hard dead lines for watering.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>